Rename app-controller.jsx for better semantics.
