+++ lib/cli/entry.js
@@ -1,4 +1,4 @@
-var ControlledApp = require('../ui/components/controlled-app.jsx')
+var ControlledApp = require('../ui/components/app-controller.jsx')
 var React = require('react')
 var ReactDom = require('react-dom')
